move HiddenPrint to its own file
