+++ python/boss_drp/utils/__init__.py
@@ -14,3 +14,4 @@ from .chpc2html import chpc2html
 from .getcard import getcard
 from .retry import retry
 from .path_to_html import path_to_html
+from .HiddenPrints import HiddenPrints
